resolve naming conflict between neo4j and knowledge_graph.neo4j
